services/message-processor/requirements.txt
pika                
psycopg2-binary        
